Move all thruster driver functions into thr module
